# Improve performance of deepseek models (Rust compiler 1.83.0+ required)